+++ lib/server/src/grpc/mod.rs
@@ -4,5 +4,11 @@
 //! Protocol v2 types to the session/buffer system.
 
 mod buffer;
+mod input;
+mod server_service;
+mod state;
 
-pub use buffer::BufferServiceImpl;
+pub use {
+    buffer::BufferServiceImpl, input::InputServiceImpl, server_service::ServerServiceImpl,
+    state::StateServiceImpl,
+};